requirements.txt
cvzone==1.5.6
ultralytics==8.0.26
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python==4.5.4.60
Pillow>=7.1.2
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
streamlit>=1.24.1
